Extract shared NemotronRNNTEngine; dedup EN + 3.5 backends
The two Nemotron backends were ~390-line clones differing only in cache
geometry, vocab/blank size, chunk length, the optional language prompt_id, and
tokenizer tag handling. Extract the shared pipeline into NemotronRNNTEngine.swift:

- NemotronRNNTConfig captures the per-model differences.
- RNNTStreamState is now a neutral top-level type (resolves the altitude nit
  where the 3.5 backend borrowed NemotronStreamingTranscriber's nested type);
  both backends typealias StreamState = RNNTStreamState, and the streaming
  protocol/controller reference RNNTStreamState directly.
- Shared free functions: nemotronMakeStreamState, nemotronTranscribeChunk
  (preprocessor → encoder(+optional prompt_id) → RNNT greedy decode),
  nemotronDecodeTokens, nemotronLoadWavAsFloats, nemotronZeroFill,
  nemotronDownloadHuggingFaceTree (with optional skipRelativePrefix).
- TranscriberError → shared NemotronRNNTError (aliased per backend).

Both transcribers are now thin wrappers owning models + config + paths. A fix
to the decode loop / mel padding / cache wiring / downloader now lands once.

Validated end-to-end: synthesized speech through the real cached multilingual
model via the shared engine returns correct punctuated text. Full suite green
(997 tests / 113 suites).
